add responsiveness to Offerings (live demo)
