+++ src/components/landing/OfferingCard.jsx
