+++ src/data/Offerings.js
@@ -1,10 +1,9 @@
-import OfferingCard from "./OfferingCard";
 import { FaUsers } from "react-icons/fa6";
 import { RiPresentationFill } from "react-icons/ri";
 import { GiPodium } from "react-icons/gi";
 import { LiaNewspaperSolid } from "react-icons/lia";
 
-const Offerings = [
+export const offerings = [
   {
     title: "SEMINARS",
     description:
@@ -30,25 +29,3 @@ const Offerings = [
     icon: <LiaNewspaperSolid className="text-5xl text-ai-blue-500" />,
   },
 ];
-
-const OfferingCards = () => {
-  return (
-    <section className="p-4">
-      <p className="text-4xl text-ai-blue-500 text-center m-8">We Offer...</p>
-      <div className="flex justify-center">
-        <div className="inline-grid grid-cols-3 gap-8 w-auto place-items-center">
-          {Offerings.map((offering, index) => (
-            <OfferingCard
-              key={index}
-              title={offering.title}
-              icon={offering.icon}
-              description={offering.description}
-            />
-          ))}
-        </div>
-      </div>
-    </section>
-  );
-};
-
-export default OfferingCards;
